pyproject.toml
[tool.poetry]
name = "hoarding-web-scout"
version = "0.1.0"
description = "An application created to watch parts of a certain college's website to know when new information is posted for all the special admissions contest. Also to save it and alert over IM bots."
authors = ["Bruno Tavares <tavares@tuta.io>"]
readme = "README.md"
packages = [{include = "hoarding_web_scout"}]

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
black = "^23.3.0"
anyio = "^3.6.2"
certifi = "^2022.12.7"
charset-normalizer = "^3.1.0"
click = "^8.1.3"
python-telegram-bot = "^20.2"
colorama = "^0.4.6"
h11 = "^0.14.0"
idna = "^3.4"
mypy-extensions = "^1.0.0"
packaging = "^23.1"
pathspec = "^0.11.1"
platformdirs = "^3.2.0"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
sniffio = "^1.3.0"
soupsieve = "^2.4.1"
urllib3 = "^1.26.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
